+++ src/main/java/roomescape/member/presentation/MemberController.java
@@ -1,19 +1,15 @@
 package roomescape.member.presentation;
 
-import jakarta.servlet.http.HttpServletRequest;
 import java.util.List;
-import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.member.application.service.MemberService;
-import roomescape.member.infrastructure.CookieAuthorizationExtractor;
 import roomescape.member.presentation.dto.MemberResponse;
 import roomescape.member.presentation.dto.SignUpRequest;
 import roomescape.member.presentation.dto.SignUpResponse;
-import roomescape.member.presentation.dto.TokenRequest;
 
 @RestController
 public class MemberController {
